+++ src/main/java/fr/alasdiablo/janoeo/holder/ItemHolder.java
@@ -17,6 +17,14 @@ public class ItemHolder {
     public static Item COPPER_INGOT;
     @ObjectHolder(RegistryHolder.MODID + ":" + RegistryHolder.TIN_INGOT)
     public static Item TIN_INGOT;
+    @ObjectHolder(RegistryHolder.MODID + ":" + RegistryHolder.ALUMINIUM_INGOT)
+    public static Item ALUMINIUM_INGOT;
+    @ObjectHolder(RegistryHolder.MODID + ":" + RegistryHolder.LEAD_INGOT)
+    public static Item LEAD_INGOT;
+    @ObjectHolder(RegistryHolder.MODID + ":" + RegistryHolder.SILVER_INGOT)
+    public static Item SILVER_INGOT;
+    @ObjectHolder(RegistryHolder.MODID + ":" + RegistryHolder.URANIUM_INGOT)
+    public static Item URANIUM_INGOT;
 
     // TODO add more ingot, nugget
 
@@ -30,6 +38,10 @@ public class ItemHolder {
             Item.Properties properties = new Item.Properties().group(JANOEO.setup.janoeoItemGroup);
             event.getRegistry().register(new Item(properties).setRegistryName(RegistryHolder.COPPER_INGOT));
             event.getRegistry().register(new Item(properties).setRegistryName(RegistryHolder.TIN_INGOT));
+            event.getRegistry().register(new Item(properties).setRegistryName(RegistryHolder.ALUMINIUM_INGOT));
+            event.getRegistry().register(new Item(properties).setRegistryName(RegistryHolder.LEAD_INGOT));
+            event.getRegistry().register(new Item(properties).setRegistryName(RegistryHolder.SILVER_INGOT));
+            event.getRegistry().register(new Item(properties).setRegistryName(RegistryHolder.URANIUM_INGOT));
         }
     }
 }